chore: move getUserSession function to auth actions file
