+++ src/libs/client/src/components/layouts/navigation-bar/nav-menu-mobile-wrapper.tsx
@@ -1,5 +1,5 @@
 import { type NavLinkArr } from "./nav-bar";
-import { getUserSession } from "@/root/src/libs/server/src/auth/utils";
+import { getUserSession } from "@/server/actions";
 import { NavMenuMobile } from "./nav-menu-mobile";
 
 export const NavMenuMobileWrapper = async function ({ navLinkArr }: { navLinkArr: NavLinkArr }) {
