+++ src/libs/server/src/hono/middleware.ts
@@ -1,6 +1,6 @@
 import { createMiddleware } from "hono/factory";
 import { Environment } from "@/root/bindings";
-import { getUserSession } from "@/server/auth/utils";
+import { getUserSession } from "@/server/actions";
 
 export const authMiddleware = createMiddleware<Environment>(async (c, next) => {
   const userSessionRes = await getUserSession();
